Feature: restructured how notifications are handled in the import/export tab
The success, warning and error notifications are now handled using the message GET parameter. This works better with the WordPress core as the parameter is removed from the URL so refreshing the page does not load the message again. The HTML and interaction in the notifications is now generated by WordPress core as well.

The messages are now stored in a static array  for better organization and ability to add or remove messages as required by the plugin.